@@ -9,8 +9,8 @@
     "esp-js": "^7.0.2",
     "esp-js-react": "^7.0.2",
     "immutable": "^4.0.0-rc.12",
-    "react": "^16.14.0",
-    "react-dom": "^16.14.0"
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1"
   },
   "devDependencies": {
     "jest": "^29.7.0",
