@@ -11,8 +11,8 @@
     "esp-js-polimer": "^7.0.2",
     "esp-js-react": "^7.0.2",
     "esp-js-ui": "^7.0.2",
-    "react": "^16.14.0",
-    "react-dom": "^16.14.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "rxjs": "^6.6.6",
     "uuid": "^8.3.2"
   },
