@@ -27,7 +27,8 @@
     "esp-js": "^7.0.2",
     "esp-js-polimer": "^7.0.2",
     "prop-types": "^15.7.2",
-    "react": "^16.14.0"
+    "react": "^18.3.1",
+    "use-sync-external-store": "^1.2.2"
   },
   "devDependencies": {
     "jest": "^29.7.0",
@@ -40,7 +41,7 @@
     "esp-js": "^7.0.0",
     "esp-js-polimer": "^7.0.0",
     "prop-types": "^15.0.0",
-    "react": "^16.0.0"
+    "react": "^18.0.0"
   },
   "gitHead": "f6791a488f3fc5f59f20e19445b848d5e2f415c4"
 }
