@@ -31,7 +31,7 @@
     "esp-js-react": "^7.0.2",
     "esp-js-rx": "^7.0.2",
     "query-string": "6.13.1",
-    "react": "^16.14.0",
+    "react": "^18.3.1",
     "rxjs": "^6.6.6",
     "uuid": "^8.3.2"
   },
@@ -41,7 +41,7 @@
     "esp-js-di": "^7.0.0",
     "esp-js-react": "^7.0.0",
     "esp-js-rx": "^7.0.0",
-    "react": "^16.0.0",
+    "react": "^18.0.0",
     "rxjs": "^6.0.0"
   }
 }
