@@ -2,11 +2,11 @@ import * as classnames from 'classnames';
 import * as React from 'react';
 import './tileContainerView.css';
 
-export type TileContainerViewProps = {
+export type TileContainerViewProps = React.PropsWithChildren<{
     title: string;
     modelId: string;
     classNames?: any;
-};
+}>;
 
 export const TileContainerView: React.FC<TileContainerViewProps> = ({title, modelId, classNames, children}) => {
     return (
