@@ -20,39 +20,55 @@
 // import for side effects
 import './polimer/polimerExtentsions';
 
-export {RouterProvider, RouterContext, useRouter} from './routerProvider';
-export {SmartComponent} from './smartComponent';
 export {ViewBinder} from './viewBinder';
+export {
+    ViewBinderConnectableProps,
+    ViewBinderConnectable
+} from './viewBinderConnectable';
 export {viewBinding, DEFAULT_VIEW_KEY} from './viewBindingDecorator';
-export * from './getEspReactRenderModel';
-export {shouldUpdateMixin} from './shouldUpdateMixin';
+export {
+    GetEspReactRenderModelMetadata,
+    GetEspReactRenderModelConsts,
+    getEspReactRenderModel
+} from './getEspReactRenderModel';
 export {createViewForModel} from './viewBindingDecorator';
 export {
-    ConnectableComponent,
     connect,
-    ConnectableComponentProps,
-    MapModelToProps,
+    ConnectableComponentFactory
+} from './connect';
+export {
     CreatePublishEventProps,
-    ConnectableView,
+    MapModelToProps,
+    ConnectableComponentProps,
     ConnectableComponentChildProps,
+    ConnectableComponent,
 } from './connectableComponent';
+export {
+    useModelSelector,
+    ConnectEqualityFn,
+    defaultConnectEqualityFn
+} from './useModelSelector';
+export {
+    RouterProvider,
+    RouterProviderProps,
+    EspRouterContext,
+    EspRouterContextProps,
+    RouterContext,
+    useRouter,
+    PublishEventDelegate,
+    PublishEventContext,
+    usePublishEvent,
+} from './espRouterContext';
 export {
     useGetModelId,
     GetModelIdDelegate,
     GetModelIdContext,
-
-    PublishEventDelegate,
-    PublishEventContext,
-    usePublishEvent,
-
     PublishModelEventDelegate,
     PublishModelEventContext,
     usePublishModelEvent,
-
     PublishModelEventWithEntityKeyDelegate,
     PublishModelEventWithEntityKeyContext,
     usePublishModelEventWithEntityKey,
-
     EspModelContext,
-    EspModelContextProps
+    EspModelContextProps,
 } from './espModelContext';
