@@ -16,33 +16,26 @@
  */
 // notice_end
 
-import * as esp from 'esp-js';
 import * as React from 'react';
-import * as PropTypes from 'prop-types';
-import { createViewForModel } from './viewBindingDecorator';
+import {createViewForModel} from './viewBindingDecorator';
+import {useRouter} from './espRouterContext';
 
 export interface ViewBinderProps {
     model: any;
     viewContext: string;
 }
 
-export class ViewBinder extends React.Component<ViewBinderProps> {
-
-    static contextTypes = {
-        router: PropTypes.instanceOf(esp.Router).isRequired
-    };
-
-    render() {
-        if(this.props.model) {
-            let {model, viewContext, ...other} = this.props;
-            let newProps = Object.assign({}, { model: this.props.model, router:this.context.router }, other);
-            return createViewForModel(
-                this.props.model,
-                newProps,
-                this.props.viewContext,
-                null
-            );
-        }
-        return null;
+export const ViewBinder: React.FC<ViewBinderProps> = (props) => {
+    if (props.model) {
+        const {model, viewContext, ...other} = props;
+        const router = useRouter();
+        const newProps = Object.assign({}, {model: props.model, router: router}, other);
+        return createViewForModel(
+            props.model,
+            newProps,
+            props.viewContext,
+            null
+        );
     }
-}
+    return null;
+};
